requirements/dev.txt
-r base.txt
django-debug-toolbar==1.8
flake8==3.4.1
ipdb==0.10.3
ipython==6.2.1
mixer==5.6.6
pytest-cov==2.4.0
pytest-django==3.1.2
pytest==3.2.3
